+++ packages/gamut-system/docs/gatsby-browser.js
@@ -1,2 +1,1 @@
-export { wrapRootElement } from './src/base';
-export { wrapPageElement } from './src/base';
+export { wrapPageElement, wrapRootElement } from './src/base';
