+++ packages/gamut-system/docs/gatsby-ssr.js
@@ -1,2 +1,1 @@
-export { wrapRootElement } from './src/base';
-export { wrapPageElement } from './src/base';
+export { wrapRootElement, wrapPageElement } from './src/base';
